ios: fix duplicate Vector2 enum and Color Float->Double widening
- Remove duplicate KANAMA_IOS_VARIANT_TYPE_VECTOR2 = 5 in variant type enum
- Add .toDouble() back on Color.r/g/b/a in tweenProperty dispatch
  (Kotlin/Native does not auto-widen Float->Double at native boundary)

+++ ios-runtime/src/iosMain/kotlin/net/multigesture/kanama/api/IosGodotApi.kt
@@ -514,7 +514,7 @@ class Tween(handle: MemorySegment) : GodotObject(handle) {
         val targetAddr = target.handle.address()
         return when (finalValue) {
             is Vector2 -> IosGodot.tweenTweenPropertyVector2(addr, targetAddr, property, finalValue.x.toDouble(), finalValue.y.toDouble(), duration)
-            is Color -> IosGodot.tweenTweenPropertyColor(addr, targetAddr, property, finalValue.r, finalValue.g, finalValue.b, finalValue.a, duration)
+            is Color -> IosGodot.tweenTweenPropertyColor(addr, targetAddr, property, finalValue.r.toDouble(), finalValue.g.toDouble(), finalValue.b.toDouble(), finalValue.a.toDouble(), duration)
             else -> IosGodot.tweenTweenPropertyVector2(addr, targetAddr, property, 0.0, 0.0, duration)
         }.takeIf { it != 0L }?.let { PropertyTweener(MemorySegment.ofAddress(it)) }
     }
